+++ src/ResultsLogger.cpp
@@ -2,9 +2,9 @@
 
 ResultsLogger::ResultsLogger() {}
 
-std::vector<float>& ResultsLogger::getSystemEnergy() { return systemEnergy; }
+std::vector<double>& ResultsLogger::getSystemEnergy() { return systemEnergy; }
 
-std::vector<float>& ResultsLogger::getIterationExecutionTime() { return iterationExecutionTime; };
+std::vector<double>& ResultsLogger::getIterationExecutionTime() { return iterationExecutionTime; };
 
 void ResultsLogger::startTimeMeasure() {
 	startTime = std::chrono::high_resolution_clock::now();
@@ -25,15 +25,15 @@ void ResultsLogger::resetLoggerData() {
 }
 
 void ResultsLogger::logSystemEnergy(std::vector<Body>& bodies) {
-	float kineticEnergy = 0.f;
-	float potentialEnergy = 0.f;
+	double kineticEnergy = 0.f;
+	double potentialEnergy = 0.f;
 
 	for(Body& b : bodies)
 		kineticEnergy += 0.5 * b.getMass() * glm::pow(glm::length(b.getVelocity()), 2);
 
 	for(int i = 0; i < bodies.size(); ++i) {
 		for(int j = i+1; j < bodies.size(); ++j) {
-			float r = glm::length(bodies[i].getPosition() - bodies[j].getPosition());
+			double r = glm::length(bodies[i].getPosition() - bodies[j].getPosition());
 			potentialEnergy += - G * bodies[i].getMass() * bodies[j].getMass() / r;
 		}
 	}
